Update: Make gui package more consistent
