+++ README.md
@@ -1,5 +1,6 @@
 <!-- Improved compatibility of back-to-top link: See: https://github.com/othneildrew/Best-README-Template/pull/73 -->
 <!-- markdownlint-disable MD033 -->
+<!-- markdownlint-disable MD041 -->
 <a name="readme-top"></a>
 <!--
 *** Thanks for checking out the Best-README-Template. If you have a suggestion*** That would make this better, please fork the repo and create a pull request
@@ -73,7 +74,7 @@
 <!-- ABOUT THE PROJECT -->
 ## About The Project
 
-Clipbird Software is a versatile tool designed to synchronize the clipboard across various devices within your local network. It supports multiple operating systems, ensuring seamless compatibility. This cross-platform software enables the effortless sharing of clipboard content, allowing users to easily copy and paste text, images, and other data between devices. With Clipbird, you can conveniently access and transfer clipboard information across different platforms, enhancing productivity and streamlining workflow in a connected environment.
+Clipbird Software Helps you to sync your clipboard data between devices connected to a local network. It is a cross-platform software that works on various platforms. It is free and open-source software that is licensed under the MIT License.
 
 <p align="right">(<a href="#readme-top">back to top</a>)</p>
 
@@ -90,7 +91,7 @@ Clipbird Software is a versatile tool designed to synchronize the clipboard acro
 <!-- GETTING STARTED -->
 ## Getting Started
 
-Clipbird is in the development stage if the project succeeds in the future, we will release the binaries to use. However you can build the project from source.
+Clipbird is in the development stage if the project succeeds in the future, we will release the binaries to use. However, you can build the project from the source.
 
 ### Prerequisites
 
@@ -171,12 +172,6 @@ Project Link: [https://github.com/srilakshmikanthanp/clipbirdesk](https://github
 
 <!-- ACKNOWLEDGMENTS -->
 ## Acknowledgments
-
-* ECM from KDE
-* C++
-* Qt6
-* OpenSSL
-* KDNSSD
 
 * [ECM from KDE](https://api.kde.org/ecm/)
 * [OpenSSL](https://www.openssl.org/)
