+++ ui/gui/window/window.cpp
@@ -5,6 +5,106 @@
 
 #include "window.hpp"
 
+namespace srilakshmikanthanp::clipbirdesk::ui::gui::internal {
+HostsList::HostsList(QWidget* parent) : QWidget(parent) {
+  // set the root layout for this widget
+  this->setLayout(verticalLayout);
+
+  // set the style
+  this->setStyleSheet(style);
+}
+
+/**
+ * @brief Set the Hosts to the list
+ */
+void HostsList::setHosts(QList<components::Host::Value> hosts) {
+  // remove & add all the hosts
+  removeAllHosts();
+  for (auto host : hosts) addHost(host);
+}
+
+/**
+ * @brief Get the All Hosts from the list
+ */
+QList<components::Host::Value> HostsList::getAllHosts() {
+  // cast the widget to host view
+  #define CAST(x) dynamic_cast<components::Host*>(x)
+
+  // create a list of hosts
+  QList<components::Host::Value> hosts;
+
+  // iterate over the layout
+  for (int i = 0; i < verticalLayout->count(); i++) {
+    auto currItem = verticalLayout->itemAt(i)->widget();
+    auto hostView = CAST(currItem);
+    hosts.append(hostView->getHost());
+  }
+
+  // return the list of hosts
+  return hosts;
+
+  // undefine the cast macro
+  #undef CAST
+}
+
+/**
+ * @brief Remove all Hosts from the list
+ */
+void HostsList::removeAllHosts() {
+  // cast the widget to host view
+  #define CAST(x) dynamic_cast<components::Host*>(x)
+
+  // iterate over the layout
+  for (int i = 0; i < verticalLayout->count(); i++) {
+    auto currItem = verticalLayout->itemAt(i)->widget();
+    auto hostView = CAST(currItem);
+    verticalLayout->removeWidget(hostView);
+  }
+
+  // undefine the cast macro
+  #undef CAST
+}
+
+/**
+ * @brief Add Host to the list
+ */
+void HostsList::addHost(components::Host::Value host) {
+  // create a new host view
+  auto hostView = new components::Host(this);
+
+  // set the host
+  hostView->setHost(host);
+
+  // slot
+  const auto slot = [&]() { emit onAction(host); };
+
+  // connect the host view signal to this signal
+  QObject::connect(hostView, &components::Host::onAction, slot);
+
+  // add the host view to the layout
+  verticalLayout->addWidget(hostView);
+}
+
+/**
+ * @brief Remove a Host from the list
+ */
+void HostsList::removeHost(components::Host::Value host) {
+  // cast the widget to host view
+  #define CAST(x) dynamic_cast<components::Host*>(x)
+
+  // iterate over the layout
+  for (int i = 0; i < verticalLayout->count(); i++) {
+    auto currItem = verticalLayout->itemAt(i)->widget();
+    auto hostView = CAST(currItem);
+    if (hostView->getHost() != host) continue;
+    verticalLayout->removeWidget(hostView);
+  }
+
+  // undefine the cast macro
+  #undef CAST
+}
+}  // namespace srilakshmikanthanp::clipbirdesk::ui::gui::internal
+
 namespace srilakshmikanthanp::clipbirdesk::ui::gui {
 //------------------------------ slots for Window -------------------------//
 
@@ -212,8 +312,8 @@ Window::Window(Window::ClipBird* controller, QWidget* parent)
   root->addWidget(tab);
 
   // using some components
-  using components::HostsList;
   using components::Tab;
+  using internal::HostsList;
 
   // connect tab changed signal
   connect(tab, &Tab::currentChanged, [&](int index) {
